Update main.py
Added a new coin-tossing functionality where the system imports the random module and then picks a random number, either 1 or 2. These correspond to heads/tails, and as no one brings a coin to matches anymore, this allows the scoring app to pick which team is batting first completely randomly. New variables were also created to enable the system to know who batted first and who needs to bat second, whether that is home, then away, or away, then home.

Signed-off-by: AlexCami1902 <alexiocam07@gmail.com>