Update main.py
Made changes after testing the program to ensure that the winner declaration says that the correct team won after the added coin toss feature. The old code just had the home team batting first, and as such, if the team batting first won, then it knew it was the home team. Now, by passing the variable through the shared.py file, it can tell who won even with the coin flip functionality.

Signed-off-by: AlexCami1902 <alexiocam07@gmail.com>