Update inputs.py
Made changes based on feeback from #33 suggesting that the text should be removed in the text box in favour of a placeholder that way the user does not need to delete the text each time they enter the teams in the app. Additional minor colour changes and button changes were also made.

Signed-off-by: AlexCami1902 <alexiocam07@gmail.com>