Update inputs.py
Added data validation to all of the inputs for the team selection and match location that the teams and location cannot be blank; this makes sure that either way, some aspect of a team is entered. This also helps with naming later on, as in theory the teams should have different names.

Signed-off-by: AlexCami1902 <alexiocam07@gmail.com>