Update main.py
Added time functionality so that the user can see what the start time of the match was and what the current time is. This allows for the user to have a much easier time using fullscreen and allows for easier tracking of how the match is going especially when games are time limited.

Signed-off-by: AlexCami1902 <alexiocam07@gmail.com>